refactor: move middlewares to a dedicated folder
