+++ packages/vike-server/src/runtime/vike-handler.ts
@@ -1,13 +1,7 @@
-import type { IncomingMessage, ServerResponse } from 'node:http'
-import compressMiddlewareFactory from '@universal-middleware/compress'
-import type { Get, RuntimeAdapter, UniversalHandler, UniversalMiddleware } from '@universal-middleware/core'
-import { getGlobalContextAsync, renderPage as _renderPage } from 'vike/server'
-import { isVercel } from '../utils/isVercel.js'
-import { connectToWeb } from './adapters/connectToWeb.js'
-import { globalStore } from './globalStore.js'
-import type { ConnectMiddleware, VikeHttpResponse, VikeOptions } from './types.js'
+import { enhance, type Get, type RuntimeAdapter, type UniversalHandler } from '@universal-middleware/core'
+import { renderPage as _renderPage } from 'vike/server'
+import type { VikeHttpResponse, VikeOptions } from './types.js'
 import { parseHeaders } from './utils/header-utils.js'
-import { assert } from '../utils/assert.js'
 
 async function renderPage<T extends RuntimeAdapter>({
   url,
@@ -42,92 +36,27 @@ async function renderPage<T extends RuntimeAdapter>({
   return pageContext.httpResponse
 }
 
-export const compressMiddleware = ((options?) => async (request, _context) => {
-  const compressionType = options?.compress ?? !isVercel()
-  const compressMiddlewareInternal = compressMiddlewareFactory()(request)
+export const renderPageHandler = ((options?) =>
+  enhance(
+    async (request, context, runtime) => {
+      const pageContextInit = { ...context, runtime, urlOriginal: request.url, headersOriginal: request.headers }
+      const response = await renderPage({
+        url: request.url,
+        headers: parseHeaders(request.headers),
+        runtimeRequest: runtime,
+        options: {
+          ...options,
+          pageContextUniversal: pageContextInit,
+          pageContext: options?.pageContext
+        }
+      })
 
-  return async (response) => {
-    if (!globalStore.isDev) {
-      const isAsset = new URL(request.url).pathname.startsWith('/assets/')
-      const shouldCompressResponse = compressionType === true || (compressionType === 'static' && isAsset)
-      if (shouldCompressResponse) {
-        return compressMiddlewareInternal(response)
-      }
+      return new Response(response.getReadableWebStream(), {
+        status: response.statusCode,
+        headers: response.headers
+      })
+    },
+    {
+      order: 0
     }
-    return response
-  }
-}) satisfies Get<[options: VikeOptions], UniversalMiddleware>
-
-export const renderPageHandler = ((options?) => async (request, context, runtime) => {
-  const nodeReq: IncomingMessage | undefined = 'req' in runtime ? runtime.req : undefined
-  let staticConfig: false | { root: string; cache: boolean } = false
-  let staticMiddleware: ConnectMiddleware | undefined
-
-  if (nodeReq) {
-    const needsUpgrade = globalStore.setupHMRProxy(nodeReq)
-
-    if (needsUpgrade) {
-      // Early response for HTTP connection upgrade
-      return new Response(null)
-    }
-
-    const { resolveStaticConfig } = await import('./utils/resolve-static-config.js')
-    staticConfig = resolveStaticConfig(options?.static)
-  }
-
-  if (globalStore.isDev) {
-    const { handleViteDevServer } = await import('./adapters/handleViteDevServer.js')
-    const handled = await connectToWeb(handleViteDevServer)(request)
-    if (handled) return handled
-  } else if (nodeReq) {
-    if (staticConfig) {
-      const handled = await connectToWeb(serveStaticFiles)(request)
-      if (handled) return handled
-    }
-  }
-
-  async function serveStaticFiles(req: IncomingMessage, res: ServerResponse): Promise<boolean> {
-    await removeBaseUrl(req)
-
-    if (!staticMiddleware) {
-      const { default: sirv } = await import('sirv')
-      staticMiddleware = sirv((staticConfig as { root: string; cache: boolean }).root, { etag: true })
-    }
-
-    return new Promise<boolean>((resolve) => {
-      res.once('close', () => resolve(true))
-      // biome-ignore lint/style/noNonNullAssertion: <explanation>
-      staticMiddleware!(req, res, () => resolve(false))
-    })
-  }
-
-  const pageContextInit = { ...context, runtime, urlOriginal: request.url, headersOriginal: request.headers }
-  const response = await renderPage({
-    url: request.url,
-    headers: parseHeaders(request.headers),
-    runtimeRequest: runtime,
-    options: {
-      ...options,
-      pageContextUniversal: pageContextInit,
-      pageContext: options?.pageContext
-    }
-  })
-
-  return new Response(response.getReadableWebStream(), {
-    status: response.statusCode,
-    headers: response.headers
-  })
-}) satisfies Get<[options: VikeOptions], UniversalHandler>
-
-async function removeBaseUrl(req: IncomingMessage) {
-  if (!req.url) return
-  const globalContext = await getGlobalContextAsync(!globalStore.isDev)
-  const baseAssets = globalContext.baseAssets as string
-  // Don't choke on older Vike versions
-  if (baseAssets === undefined) return
-  const { url } = req
-  assert(url.startsWith('/'))
-  let urlWithoutBase = url.slice(baseAssets.length)
-  if (!urlWithoutBase.startsWith('/')) urlWithoutBase = `/${urlWithoutBase}`
-  req.url = urlWithoutBase
-}
+  )) satisfies Get<[options: VikeOptions], UniversalHandler>
