+++ packages/vike-server/src/vike.handler.ts
@@ -1,6 +1,7 @@
-import { type Get, type RuntimeAdapter, type UniversalHandler, pipe } from '@universal-middleware/core'
+import { type Get, pipe, type RuntimeAdapter, type UniversalHandler } from '@universal-middleware/core'
 import type { VikeOptions } from './runtime/types.js'
-import { compressMiddleware, renderPageHandler } from './runtime/vike-handler.js'
+import { renderPageHandler } from './runtime/vike-handler.js'
+import { globalStore } from './runtime/globalStore.js'
 
 // https://vike.dev/pageContext#typescript
 declare global {
@@ -11,9 +12,16 @@ declare global {
   }
 }
 
-const renderPageUniversal = ((options?) => pipe(compressMiddleware(options), renderPageHandler(options))) satisfies Get<
-  [options: VikeOptions],
-  UniversalHandler
->
+let renderPageUniversal: Get<[options: VikeOptions], UniversalHandler>
+
+if (globalStore.isDev) {
+  const { devServerMiddleware } = await import('./middlewares/devServer.js')
+  renderPageUniversal = (options?) => pipe(devServerMiddleware(), renderPageHandler(options))
+} else {
+  const { compressMiddleware } = await import('./middlewares/compress.js')
+  const { serveStaticMiddleware } = await import('./middlewares/serveStatic.js')
+  renderPageUniversal = (options?) =>
+    pipe(compressMiddleware(options), serveStaticMiddleware(options), renderPageHandler(options))
+}
 
 export default renderPageUniversal
