+++ packages/vike-server/tsconfig.json
@@ -3,7 +3,8 @@
     "rootDir": "./src",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
-    "target": "ES2019",
+    "target": "ES2022",
+    "verbatimModuleSyntax": true,
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "types": ["vite/client"],
     "esModuleInterop": true,
